pyproject.toml
[tool.poetry]
name = "r_website_backend"
version = "0.1.0"
description = "Backend for my personal/blog website. Will serve up blog posts, blog post previews, etc."
authors = ["Richard Hausman <hausman.richard@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.1.1"
djangorestframework = "^3.11.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
